fix migration

+++ database/migrations/2022_08_14_091610_remove_old_tags.php
@@ -16,6 +16,9 @@ class RemoveOldTags extends Migration
     public function up()
     {
         $tag = Tag::findFromString('awaiting_webhook_publish');
-        DB::table('taggables')->where('tag_id', $tag->getKey())->delete();
+
+        if ($tag) {
+            DB::table('taggables')->where('tag_id', $tag->getKey())->delete();
+        }
     }
 }
